refactor(docs): time-input dx
