+++ apps/docs/content/components/time-input/error-message-function.raw.jsx
@@ -1,0 +1,15 @@
+import {TimeInput} from "@nextui-org/react";
+
+export default function App() {
+  return (
+    <TimeInput
+      isInvalid
+      errorMessage={(value) => {
+        if (value.isInvalid) {
+          return "Please enter a valid time";
+        }
+      }}
+      label="Event Time"
+    />
+  );
+}
