+++ apps/docs/content/components/time-input/error-message.raw.jsx
@@ -1,0 +1,5 @@
+import {TimeInput} from "@nextui-org/react";
+
+export default function App() {
+  return <TimeInput isInvalid errorMessage="Please enter a valid time" label="Event Time" />;
+}
