+++ apps/docs/content/components/time-input/hide-timezone.ts
@@ -1,15 +1,4 @@
-const App = `import {TimeInput} from "@nextui-org/react";
-import {parseZonedDateTime} from "@internationalized/date";
-
-export default function App() {
-  return (
-    <TimeInput 
-      label="Meeting time"
-      hideTimeZone 
-      defaultValue={parseZonedDateTime("2022-11-07T10:45[America/Los_Angeles]")} 
-    />
-  );
-}`;
+import App from "./hide-timezone.raw.jsx?raw";
 
 const react = {
   "/App.jsx": App,
