+++ apps/docs/content/components/time-input/hour-cycle.ts
@@ -1,16 +1,4 @@
-const App = `import {TimeInput} from "@nextui-org/react";
-import {Time, parseZonedDateTime} from "@internationalized/date";
-
-export default function App() {
-  return (
-    <TimeInput 
-      label="Meeting time"
-      hourCycle={24}
-      defaultValue={parseZonedDateTime("2022-11-07T00:45[America/Los_Angeles]")}
-      granularity="minute" 
-    />
-  );
-}`;
+import App from "./hour-cycle.raw.jsx?raw";
 
 const react = {
   "/App.jsx": App,
