+++ apps/docs/content/components/time-input/label-placement.raw.jsx
@@ -1,0 +1,27 @@
+import {TimeInput} from "@nextui-org/react";
+import {Time} from "@internationalized/date";
+
+export default function App() {
+  return (
+    <div className="w-full max-w-xl flex flex-col items-end gap-4">
+      <TimeInput
+        defaultValue={new Time(11, 45)}
+        description="inside"
+        label="Event Time"
+        labelPlacement="inside"
+      />
+      <TimeInput
+        defaultValue={new Time(11, 45)}
+        description="outside"
+        label="Event Time"
+        labelPlacement="outside"
+      />
+      <TimeInput
+        defaultValue={new Time(11, 45)}
+        description="outside-left"
+        label="Event Time"
+        labelPlacement="outside-left"
+      />
+    </div>
+  );
+}
