+++ apps/docs/content/components/time-input/max-time-value.raw.jsx
@@ -1,0 +1,6 @@
+import {TimeInput} from "@nextui-org/react";
+import {Time} from "@internationalized/date";
+
+export default function App() {
+  return <TimeInput defaultValue={new Time(18)} maxValue={new Time(17)} />;
+}
