+++ apps/docs/content/components/time-input/max-time-value.ts
@@ -1,14 +1,4 @@
-const App = `import {TimeInput} from "@nextui-org/react";
-import {Time} from "@internationalized/date";
-
-export default function App() {
-  return (
-    <TimeInput 
-      defaultValue={new Time(18)}
-      maxValue={new Time(17)}
-    />
-  );
-}`;
+import App from "./max-time-value.raw.jsx?raw";
 
 const react = {
   "/App.jsx": App,
