+++ apps/docs/content/components/time-input/min-time-value.ts
@@ -1,14 +1,4 @@
-const App = `import {TimeInput} from "@nextui-org/react";
-import {Time} from "@internationalized/date";
-
-export default function App() {
-  return (
-    <TimeInput 
-      defaultValue={new Time(8)}
-      minValue={new Time(9)}
-    />
-  );
-}`;
+import App from "./min-time-value.raw.jsx?raw";
 
 const react = {
   "/App.jsx": App,
