+++ apps/docs/content/components/time-input/placeholder-value.raw.jsx
@@ -1,0 +1,6 @@
+import {TimeInput} from "@nextui-org/react";
+import {Time} from "@internationalized/date";
+
+export default function App() {
+  return <TimeInput label="Meeting time" placeholderValue={new Time(9)} />;
+}
