+++ apps/docs/content/components/time-input/placeholder-value.ts
@@ -1,14 +1,4 @@
-const App = `import {TimeInput} from "@nextui-org/react";
-import {Time} from "@internationalized/date";
-
-export default function App() {
-  return (
-    <TimeInput 
-      label="Meeting time"
-      placeholderValue={new Time(9)}
-    />
-  );
-}`;
+import App from "./placeholder-value.raw.jsx?raw";
 
 const react = {
   "/App.jsx": App,
