+++ apps/docs/content/components/time-input/read-only.ts
@@ -1,15 +1,4 @@
-const App = `import {TimeInput} from "@nextui-org/react";
-import {Time} from "@internationalized/date";
-
-export default function App() {
-  return (
-    <TimeInput 
-      isReadOnly 
-      label="Event Time" 
-      defaultValue={new Time(11, 45)} 
-    />
-  );
-}`;
+import App from "./read-only.raw.jsx?raw";
 
 const react = {
   "/App.jsx": App,
