+++ apps/docs/content/components/time-input/required.raw.jsx
@@ -1,0 +1,5 @@
+import {TimeInput} from "@nextui-org/react";
+
+export default function App() {
+  return <TimeInput isRequired label="Event Time" />;
+}
