+++ apps/docs/content/components/time-input/usage.ts
@@ -1,14 +1,4 @@
-const App = `import {TimeInput} from "@nextui-org/react";
-import {Time} from "@internationalized/date";
-
-export default function App() {
-  return (
-    <div className="flex flex-wrap gap-4">
-      <TimeInput label="Event Time" />
-      <TimeInput label="Event Time" defaultValue={new Time(11, 45)} />
-    </div>
-  );
-}`;
+import App from "./usage.raw.jsx?raw";
 
 const react = {
   "/App.jsx": App,
