+++ apps/docs/content/components/time-input/with-description.raw.jsx
@@ -1,0 +1,5 @@
+import {TimeInput} from "@nextui-org/react";
+
+export default function App() {
+  return <TimeInput description="Please enter your meeting time" label="Event Time" />;
+}
