+++ apps/docs/content/components/time-input/without-label.raw.jsx
@@ -1,0 +1,6 @@
+import {TimeInput} from "@nextui-org/react";
+import {Time} from "@internationalized/date";
+
+export default function App() {
+  return <TimeInput defaultValue={new Time(11, 45)} label={null} />;
+}
